Attribute Macro sewup_derive::ewasm_fn

#[ewasm_fn]

helps you to build your handlers in the contract

This macro also generate the function signature, you can use ewasm_fn_sig! macro to get your function signature;

#[ewasm_fn]
fn check_input_object(s: SimpleStruct) -> anyhow::Result<()> {
    Ok(())
}

#[ewasm_main]
fn main() -> Result<()> {
    let contract = Contract::new()?;
    match contract.get_function_selector()? {
        ewasm_fn_sig!(check_input_object) => ewasm_input_from!(contract move check_input_object)?,
        _ => return Err(Error::UnknownHandle.into()),
    };
    Ok(())
}

There are two kind of inputs for ewasm_fn macro, first is functional signature, second and more are the fields of abijson. The functional signature can be specific as following ways. #[ewasm_fn(00fdd58e)] or #[ewasm_fn(“00fdd58e”)]

constant=true,
inputs=[
    { "internalType": "address", "name": "account", "type": "address" },
    { "internalType": "uinit256", "name": "token_id", "type": "uinit256" }
],
name=balanceOf,
outputs=[
    { "internalType": "uint256", "name": "", "type": "uint256" }
],
payable=false,
stateMutability=view
)]

The fields are not required, it can use default value if not provided. The default values of constant, payable are false; the default values of inputs and outputs are []; the default value of stateMutability is view; the default name is the camel case style of the function name.